requirements.txt
streamlit
faiss-cpu
openai
numpy
feedparser
newspaper3k
